Carrito

+++ tpo/src/main/java/com/uade/tpo/service/CarritoService.java
@@ -12,13 +12,16 @@ import com.uade.tpo.controller.request.CarritoRequest;
 import com.uade.tpo.controller.request.CarritoResponse;
 import com.uade.tpo.entity.Carrito;
 import com.uade.tpo.entity.CarritoProductos;
-import com.uade.tpo.entity.Product;
+import com.uade.tpo.entity.Producto;
 import com.uade.tpo.exception.BadProductQuantityException;
+import com.uade.tpo.exception.CartAlreadyEmptyException;
 import com.uade.tpo.exception.CartNotFoundException;
+import com.uade.tpo.exception.ExceededCartQuantityException;
 import com.uade.tpo.exception.ProductNotFoundException;
+import com.uade.tpo.exception.ProductNotInCartException;
 import com.uade.tpo.repository.CarritoProductosRepository;
 import com.uade.tpo.repository.CarritoRepository;
-import com.uade.tpo.repository.ProductoRepository;
+import com.uade.tpo.repository.ProductosRepository;
 
 import jakarta.transaction.Transactional;
 
@@ -30,35 +33,59 @@ public class CarritoService {
     private CarritoRepository carritoRepository;
     
     @Autowired
-    private ProductoRepository productoRepository;
+    private ProductosRepository productoRepository;
     
     @Autowired
     private CarritoProductosRepository carritoProdRepository;
 
     public Carrito getCarritoById(final Long carritoId) throws CartNotFoundException {
     	return this.carritoRepository.findById(carritoId)
-    			.orElseThrow(() -> new CartNotFoundException("Carrito with id " + carritoId + " not found"));
-    }
-
-    public Carrito getCarritoByUserId(final Long userId) throws CartNotFoundException {
-        return this.carritoRepository.findByUserId(userId)
-                .orElseThrow(() -> new CartNotFoundException("Carrito for user id " + userId + " not found"));
+    			.orElseThrow(() -> new CartNotFoundException("Carrito con id " + carritoId + " no encontrado"));
     }
 
     @Transactional
-    public ResponseEntity<CarritoResponse> addToCarrito(final Long carritoId, final CarritoRequest request) throws CartNotFoundException, ProductNotFoundException {
+    public ResponseEntity<CarritoResponse> removeFromCarrito(final Long carritoId, final CarritoRequest request)  throws CartNotFoundException, ProductNotInCartException {
+        Carrito carrito = getCarritoById(carritoId);
+        
+        Optional<CarritoProductos> optCarritoProducto = carrito.getCarritoProductos().stream()
+                .filter(cp -> cp.getProducto().getIdProductos().equals(request.getProductoId()))
+                .findFirst();
+        
+        if (optCarritoProducto.isPresent()) {
+            CarritoProductos carritoProducto = optCarritoProducto.get();
+            int cantidadEliminada = carritoProducto.getCantidad();
+
+            float nuevoTotal = carrito.getTotal() - (carritoProducto.getProducto().getPrecioConDescuento() * cantidadEliminada);
+            carrito.setTotal(nuevoTotal);
+
+            int nuevaCantidadProducto = carritoProducto.getProducto().getCantidad() + cantidadEliminada;
+            carritoProducto.getProducto().setCantidad(nuevaCantidadProducto);
+           
+            carrito.getCarritoProductos().remove(carritoProducto);
+            
+            carritoRepository.save(carrito);
+            carritoProdRepository.delete(carritoProducto);
+            
+            return ResponseEntity.ok(new CarritoResponse("Producto borrado del carrito exitosamente.", carrito));
+        } else {
+        	throw new ProductNotInCartException("El producto no se encuentra en el carrito con id: " + carritoId);
+        }
+
+    }
+    
+    @Transactional
+    public ResponseEntity<CarritoResponse> addToCarrito(final Long carritoId, final CarritoRequest request) throws CartNotFoundException, ProductNotFoundException, ExceededCartQuantityException {
 
         Carrito carrito = getCarritoById(carritoId);
         int cantidad = request.getCantidad();
 
-        Optional<Product> optProducto = productoRepository.findById(request.getProductoId());
+        Optional<Producto> optProducto = productoRepository.findById(request.getProductoId());
 
         if (optProducto.isPresent()) {
-            Product producto = optProducto.get();
+        	Producto producto = optProducto.get();
 
-            // El producto ya está en el carrito
             Optional<CarritoProductos> optCarritoProducto = carrito.getCarritoProductos().stream()
-                    .filter(cp -> cp.getProducto().getId().equals(producto.getId()))
+                    .filter(cp -> cp.getProducto().getIdProductos().equals(producto.getIdProductos()))
                     .findFirst();
 
             // Si el producto ya está en el carrito
@@ -70,14 +97,14 @@ public class CarritoService {
                     producto.setCantidad(producto.getCantidad() - cantidad);
                     carritoProducto.setCantidad(nuevaCantidad);
                     
-                    float total = carrito.getTotal() + (producto.getPrecio() * cantidad);
+                    float total = carrito.getTotal() + (producto.getPrecioConDescuento() * cantidad);
                     carrito.setTotal(total);
                     
                     productoRepository.save(producto);
                     carritoRepository.save(carrito);
                     return ResponseEntity.ok(new CarritoResponse("Cantidad de producto en el carrito actualizada.", carrito));
                 } else {
-                    return ResponseEntity.badRequest().body(new CarritoResponse("No hay suficiente cantidad del producto en el inventario.", "BAD_REQUEST"));
+                	throw new ExceededCartQuantityException("No hay suficiente stock de ese producto");
                 }
                 
             } else {
@@ -85,7 +112,7 @@ public class CarritoService {
                 if (producto.getCantidad() >= cantidad) {
                     producto.setCantidad(producto.getCantidad() - cantidad);
 
-                    float total = carrito.getTotal() + (producto.getPrecio() * cantidad);
+                    float total = carrito.getTotal() + (producto.getPrecioConDescuento() * cantidad);
                     carrito.setTotal(total);
 
                     CarritoProductos carritoProducto = new CarritoProductos();
@@ -108,24 +135,83 @@ public class CarritoService {
         }
     }
 
+    
     @Transactional
-	public ResponseEntity<CarritoResponse> vaciarCarrito(Long carritoId) throws CartNotFoundException {
+	public ResponseEntity<CarritoResponse> emptyCarrito(Long carritoId) throws CartNotFoundException, CartAlreadyEmptyException {
 		Carrito carrito = getCarritoById(carritoId);
 		
-		carrito.setTotal(0);
+		if (carrito.getCarritoProductos().isEmpty()) {
+			throw new CartAlreadyEmptyException("El carrito ya está vacio");
+		}
 		
-        Iterator<CarritoProductos> iterator = carrito.getCarritoProductos().iterator();
+		Iterator<CarritoProductos> iterator = carrito.getCarritoProductos().iterator();
         while (iterator.hasNext()) {
             CarritoProductos carritoProducto = iterator.next();
-            iterator.remove(); 
+            iterator.remove();
+
+            Producto producto = carritoProducto.getProducto();
+            
+            int cantidadActual = producto.getCantidad();
+            int cantidadEnCarrito = carritoProducto.getCantidad();
+            
+            producto.setCantidad(cantidadActual + cantidadEnCarrito);
+            
+            productoRepository.save(producto);
             carritoProdRepository.delete(carritoProducto);
         }
 
+        carrito.setTotal(0);
         
 		this.carritoRepository.save(carrito);
 		
 		return ResponseEntity.ok(new CarritoResponse("Carrito vaciado con éxito.", ""));
 	}
 
+    
+    @Transactional
+	public ResponseEntity<CarritoResponse> substractFromCarrito(Long carritoId, CarritoRequest request) throws CartNotFoundException, ProductNotFoundException, ProductNotInCartException, ExceededCartQuantityException{
 
+		Carrito carrito = getCarritoById(carritoId);
+        int cantidad = request.getCantidad();
+
+        Optional<Producto> optProducto = productoRepository.findById(request.getProductoId());
+
+        if (optProducto.isPresent()) {
+        	Producto producto = optProducto.get();
+
+            Optional<CarritoProductos> optCarritoProducto = carrito.getCarritoProductos().stream()
+                    .filter(cp -> cp.getProducto().getIdProductos().equals(producto.getIdProductos()))
+                    .findFirst();
+
+            if (optCarritoProducto.isPresent()) {
+                CarritoProductos carritoProducto = optCarritoProducto.get();
+
+                int nuevaCantidadDelCarrito = carritoProducto.getCantidad() - cantidad;
+                float total = 0.0f;
+                
+                if (nuevaCantidadDelCarrito < 0) {
+                	throw new ExceededCartQuantityException("La cantidad a eliminar excede el la cantidad del carrito.");
+                } else if (nuevaCantidadDelCarrito == 0) {
+                	carrito.getCarritoProductos().remove(carritoProducto);
+                    carritoProdRepository.delete(carritoProducto);
+                } else {
+                	carritoProducto.setCantidad(nuevaCantidadDelCarrito);
+                }
+                total = carrito.getTotal() - (producto.getPrecioConDescuento() * cantidad);
+
+                producto.setCantidad(producto.getCantidad() + cantidad);
+
+                carrito.setTotal(total);
+                
+                productoRepository.save(producto);
+                carritoRepository.save(carrito);
+                return ResponseEntity.ok(new CarritoResponse("Cantidad de producto en el carrito actualizada.", carrito));
+            } else {
+            	throw new ProductNotInCartException("El producto no se encuentra en el carrito con id: " + carritoId);
+            }
+        } else {
+            throw new ProductNotFoundException("No se encontró el producto con id: " + request.getProductoId());
+        }
+    }
+	
 }
